CRUD Update

+++ src/main/java/com/example/demo/repository/CommunityRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CommunityRepository extends JpaRepository<Community,Integer> {
+public interface CommunityRepository extends JpaRepository<Community,Long> {
 }
